removed ExtAudioConverter

+++ r128x-cli/CliController.m
@@ -31,7 +31,7 @@
 - (void) doMeasure {
     OSStatus err = noErr;
     double il, lra;
-    Float32 max_tp;
+    double max_tp;
     self.status = ExtAudioReader((__bridge CFStringRef)(self.filePath), &il, &lra, &max_tp);
     if (self.status != noErr) {
         self.il = NAN;
